Use single, specifiable, Fog connection object.
Add helper method on Vcloud::Core to return a common Fog connection
object.

Also specify a 'setter' method to allow the user to specify their own
Fog connection.

This has been created on Vcloud::Core rather than Vcloud::Fog, as
Vcloud::Fog is to be removed due to not following Gem namespace
conventions.

No tests confirming the behaviour of this as-yet.