Add a 'default Fog connection' class
We currently create many Fog connections - due to the way that we
call Fog::Compute::VcloudDirector - rather than re-using the same
connection.

We also do not allow our API users to inject their own Fog connection,
if they wish to (generally, to provide their own authentication method,
or to be able to 'juggle' connections to different Organizations for
comparison).

This new class forms the basis of a solution to this - create a connection
if we do not already have one, and then use this connection unless we are
provided with an alternative.

Note that we use an instance variable (@default_connection) in Connection
rather than a class variable. This is so that subclasses of Connection do not
automatically inherit this default connection, and they are free to set their own.

NB: In the tests for this, we must reset the Connection class instance variable
    @default_connection between tests - to do this we have to hook into the Rspec
        configuration directly.